Provide feedback on resource availability as soon as it's added to the attendees form.
If response type is Kronolith_Resource::RESPONSETYPE_AUTO then use the start and end datetime
values that are now passed from the event form to determine if the resource is free.

Still need to figure out how to (efficiently) pass the datetime values to take into account
AM/PM. Probably break out the javascript from the onClick handler into it's function and
provide logic to check the am/pm fields? Is there an easier/prettier way?