For FB events, don't send a link, but a url property instead.
Linking the event directly to the FB page is distruptive, and there
is no way to navigate directly back to Horde. Using the back button
will bring you back to horde, but will then immediately redirect back
to FB.

Just display the URL in the event view url field instead.