slight hack to ensure we don't have a driver type prefix. Fixes Bug: 10412
The iTip viewer in IMP uses fb_cals to look for possible conflicting events.
The fb calendars are returned with the driver prefix to the calendar names, but
listEvents does not expect them.

Not the cleanest fix, but can't think of a better place. Don't want the iTip
viewer to have to know about the internals of how Kronolith stores calendar names.

A related issue/limitation is that it's possible to choose external calendars as
fb sources, but when trying to obtain the list of events from those calendars via
the API, only internal events are considered since Kronolith_Api::listEvents only
returns internal calendar events.