Automatically invite organizer to his own event
Based on a survey of several calendar clients (Google Calendar, iCal, Thunderbird with Lightning, Kmail, Evolution and Horde) we (Jan, Chuck and Ben) decided that the most compatible way to show the full list of attendees was to include the organizer by default to his own events.  If the organizer does not want to participate in the event (for example, the case of an administrative assistant scheduling a meeting on behalf of someone else), then the organizer can remove himself from the attendee list.