Fix dealing with resources when adding from dynamic view.
Need to parse resources after the rest of the event form is read so
we have the time data. Also, if we are an ajax request, don't
use the session value for resources, otherwise we
end up holding a reference to the session array - and this screws up
resource acceptance for groups. (The session value is only used from the
traditional view).