The default is an unset baseid.
Since the default value for the event_baseid column is an empty string, there is a difference to setting it to null. This is important because we later retrieve only the base events when exporting to DAV/ICS by searching for event_baseid = ''.

Fix broken events with:
UPDATE kronolith_events SET event_baseid = '' WHERE event_baseid IS NULL