Don't read email addresses from Horde user attendees.
This is potential information disclosure, because people could go fishing for user email addresses that might otherwise be unknown. The real name is not as sensitive, so keep fetching these.
Since user attendees are not going to go through the iTip machinery, don't export them as ATTENDEE attributes either. Introduce X-HORDE-ATTENDEE instead, to allow export and import to and from iCalendar without losing information.