Don't add fb if attendee tab isn't visible.
Fixes a .defer() loop due to the td.getWidth() check
further down the method. This was causing FF to become
unresponsive if the event view was left open long enough
without viewing the attendee tab.

Not sure what the check for td.getWidth() is supposed to
catch/prevent...so the correct fix may be to just remove that
conditional. It looks like the insertFreeBusy method is triggered
when viewing the AttendeeTab each time anyway.