Allow returning a hash from import().
Needed since attachment addition/deletions are handled by the
message objects in activesync, instead of by a folder command
as the rest of addition/deletions are. Clients need to be able
to map their clientId with the returned server UID. Will also
allow us to return the syncstamp when the rest of the API
is refactored to allow this.